Update runtime
API is now:
	setData(qml-source-code)
	setSource(url-to-qml-source)

setSource no longer implicitly calls load(), which
means it can be called as soon as the runtime object
has been created, also in situation where kicking off
the downloading of Qt is not appropriate.

Set base url to the empty url instead of ("/")

Remove usages of "self" in qtquickruntime.js. Instance
variable lookup works, as long as we don't shadow the
variable names. ("this" is not useful here)